requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# HTTP Client
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# YAML Configuration
pyyaml>=6.0.0

# Database (SQLite is built-in, no package needed)

# Optional: For development
# Install development dependencies with: pip install -r requirements-dev.txt
# pytest>=7.4.0
# pytest-asyncio>=0.21.0
# black>=23.11.0
# isort>=5.12.0
# mypy>=1.7.0
